Refactored optimizers, broke out abs classes to base.py and added docstrings
