+++ tests/conftest.py
@@ -1,0 +1,10 @@
+import os
+import pytest
+
+# remove numba compilation for testing
+os.environ["NUMBA_DISABLE_JIT"] = 1
+
+
+
+
+
